initial block types

+++ code/Voxel/World/VoxelWorld.Collisions.cs
@@ -2,24 +2,24 @@
 
 partial class VoxelWorld
 {
-	private bool trySpreadX( Chunk chunk, bool canSpreadX, ref bool[,,] tested, (ushort x, ushort y, ushort z) start, ref (int x, int y, int z) size )
+	private bool trySpreadX( Chunk chunk, bool canSpreadX, ref bool[,,] tested, (byte x, byte y, byte z) start, ref (int x, int y, int z) size )
 	{
 		var yLimit = start.y + size.y;
 		var zLimit = start.z + size.z;
-		for ( ushort y = start.y; y < yLimit && canSpreadX; ++y )
-		for ( ushort z = start.z; z < zLimit; ++z )
+		for ( byte y = start.y; y < yLimit && canSpreadX; ++y )
+		for ( byte z = start.z; z < zLimit; ++z )
 		{
-			var newX = (ushort)(start.x + size.x);
+			var newX = (byte)(start.x + size.x);
 			if ( newX >= Chunk.DEFAULT_WIDTH || tested[newX, y, z] || chunk.GetVoxel( newX, y, z ) == null )
 				canSpreadX = false;
 		}
 
 		if ( canSpreadX )
 		{
-			for ( ushort y = start.y; y < yLimit; ++y )
-			for ( ushort z = start.z; z < zLimit; ++z )
+			for ( byte y = start.y; y < yLimit; ++y )
+			for ( byte z = start.z; z < zLimit; ++z )
 			{
-				var newX = (ushort)(start.x + size.x);
+				var newX = (byte)(start.x + size.x);
 				tested[newX, y, z] = true;
 
 				if ( chunk.GetVoxel( newX, y, z ) == null )
@@ -32,24 +32,24 @@ partial class VoxelWorld
 		return canSpreadX;
 	}
 
-	private bool trySpreadY( Chunk chunk, bool canSpreadY, ref bool[,,] tested, (ushort x, ushort y, ushort z) start, ref (int x, int y, int z) size )
+	private bool trySpreadY( Chunk chunk, bool canSpreadY, ref bool[,,] tested, (byte x, byte y, byte z) start, ref (int x, int y, int z) size )
 	{
 		var xLimit = start.x + size.x;
 		var zLimit = start.z + size.z;
-		for ( ushort x = start.x; x < xLimit && canSpreadY; ++x )
-		for ( ushort z = start.z; z < zLimit; ++z )
+		for ( byte x = start.x; x < xLimit && canSpreadY; ++x )
+		for ( byte z = start.z; z < zLimit; ++z )
 		{
-			var newY = (ushort)(start.y + size.y);
+			var newY = (byte)(start.y + size.y);
 			if ( newY >= Chunk.DEFAULT_DEPTH || tested[x, newY, z] || chunk.GetVoxel( x, newY, z ) == null )
 				canSpreadY = false;
 		}
 
 		if ( canSpreadY )
 		{
-			for ( ushort x = start.x; x < xLimit; ++x )
-			for ( ushort z = start.z; z < zLimit; ++z )
+			for ( byte x = start.x; x < xLimit; ++x )
+			for ( byte z = start.z; z < zLimit; ++z )
 			{
-				var newY = (ushort)(start.y + size.y);
+				var newY = (byte)(start.y + size.y);
 				tested[x, newY, z] = true;
 
 				if ( chunk.GetVoxel( x, newY, z ) == null )
@@ -62,24 +62,24 @@ partial class VoxelWorld
 		return canSpreadY;
 	}
 
-	private bool trySpreadZ( Chunk chunk, bool canSpreadZ, ref bool[,,] tested, (ushort x, ushort y, ushort z) start, ref (int x, int y, int z) size )
+	private bool trySpreadZ( Chunk chunk, bool canSpreadZ, ref bool[,,] tested, (byte x, byte y, byte z) start, ref (int x, int y, int z) size )
 	{
 		var xLimit = start.x + size.x;
 		var yLimit = start.y + size.y;
-		for ( ushort x = start.x; x < xLimit && canSpreadZ; ++x )
-		for ( ushort y = start.y; y < yLimit; ++y )
+		for ( byte x = start.x; x < xLimit && canSpreadZ; ++x )
+		for ( byte y = start.y; y < yLimit; ++y )
 		{
-			var newZ = (ushort)(start.z + size.z);
+			var newZ = (byte)(start.z + size.z);
 			if ( newZ >= Chunk.DEFAULT_HEIGHT || tested[x, y, newZ] || chunk.GetVoxel( x, y, newZ ) == null )
 				canSpreadZ = false;
 		}
 
 		if ( canSpreadZ )
 		{
-			for ( ushort x = start.x; x < xLimit; ++x )
-			for ( ushort y = start.y; y < yLimit; ++y )
+			for ( byte x = start.x; x < xLimit; ++x )
+			for ( byte y = start.y; y < yLimit; ++y )
 			{
-				var newZ = (ushort)(start.z + size.z);
+				var newZ = (byte)(start.z + size.z);
 				tested[x, y, newZ] = true;
 
 				if ( chunk.GetVoxel( x, y, newZ ) == null )
